UPDATE: status icons
This commit update all the status icons!
Because completing the already existing icons (whitesur) with numix type
created differences within the same type of icons (ex: battery 10% with numix
style and battery 20% with whitesur style) I decided to update ALL the icons
to a Numix style and update the colors of all of them to use Nord colors.